feat: add tool and nodes sidebars, implement highlight text utility, and enhance message node selection
