+++ app/flow/components/root-node.tsx
@@ -1,8 +1,8 @@
 import { Card, CardHeader, CardTitle } from "@/components/ui/card";
-import { Handle, Position } from "@xyflow/react";
+import { Handle, NodeProps, Position } from "@xyflow/react";
 import { SendHorizonal } from "lucide-react";
 
-const RootNode = ({ id }: { id: string }) => {
+const RootNode = ({ id }: NodeProps) => {
   return (
     <Card className="w-[250px]">
       <CardHeader className="flex flex-row gap-2 p-3">
